@@ -43,7 +43,7 @@
   },
   "homepage": "https://github.com/jumpinjackie/mapguide-react-layout#readme",
   "dependencies": {
-    "@blueprintjs/core": "1.35.7",
+    "@blueprintjs/core": "1.36.0",
     "history": "4.7.2",
     "ismobilejs": "0.4.1",
     "lodash.debounce": "4.0.8",
@@ -53,11 +53,11 @@
     "proj4": "2.4.4",
     "prop-types": "15.6.1",
     "query-string": "5.1.1",
-    "react": "^16.3.0",
+    "react": "^16.3.2",
     "react-addons-css-transition-group": "15.6.2",
     "react-addons-test-utils": "15.6.2",
     "react-addons-update": "15.6.2",
-    "react-dom": "16.3.0",
+    "react-dom": "16.3.2",
     "react-draggable": "3.0.5",
     "react-measure": "1.4.7",
     "react-redux": "5.0.7",
@@ -69,23 +69,23 @@
     "redux-thunk": "2.2.0",
     "shortid": "2.2.8",
     "tslib": "1.9.0",
-    "url-parse": "1.2.0",
+    "url-parse": "1.4.0",
     "whatwg-fetch": "2.0.4"
   },
   "devDependencies": {
-    "@types/enzyme": "3.1.9",
-    "@types/jest": "22.2.2",
+    "@types/enzyme": "3.1.10",
+    "@types/jest": "22.2.3",
     "@types/lodash.debounce": "4.0.3",
     "@types/lodash.uniq": "4.5.3",
     "@types/lodash.xor": "4.5.3",
     "@types/prop-types": "15.5.2",
     "@types/query-string": "5.1.0",
-    "@types/react": "16.3.0",
+    "@types/react": "16.3.11",
     "@types/react-addons-css-transition-group": "15.0.4",
     "@types/react-addons-update": "0.14.19",
-    "@types/react-dom": "16.0.4",
+    "@types/react-dom": "16.0.5",
     "@types/react-measure": "0.4.8",
-    "@types/react-redux": "^5.0.15",
+    "@types/react-redux": "^5.0.18",
     "@types/shortid": "0.0.29",
     "bestzip": "1.1.4",
     "cache-loader": "^1.2.2",
@@ -99,23 +99,23 @@
     "fork-ts-checker-webpack-plugin": "^0.4.1",
     "jest": "22.4.3",
     "jsonfile": "4.0.0",
-    "less": "3.0.1",
+    "less": "3.0.2",
     "less-loader": "4.1.0",
     "raf": "^3.4.0",
     "raw-loader": "0.5.1",
-    "react-test-renderer": "16.3.0",
+    "react-test-renderer": "16.3.2",
     "rimraf": "2.6.2",
     "source-map-loader": "0.2.3",
-    "style-loader": "0.20.3",
+    "style-loader": "0.21.0",
     "thread-loader": "^1.1.5",
-    "ts-jest": "22.4.2",
-    "ts-loader": "4.1.0",
+    "ts-jest": "22.4.4",
+    "ts-loader": "4.2.0",
     "tslint": "5.9.1",
     "typedoc": "0.11.1",
-    "typescript": "2.8.1",
+    "typescript": "2.8.3",
     "url-loader": "1.0.1",
-    "webpack": "4.4.1",
-    "webpack-cli": "^2.0.13"
+    "webpack": "4.6.0",
+    "webpack-cli": "^2.0.15"
   },
   "greenkeeper": {
     "ignore": [
