@@ -24,7 +24,8 @@ export class ManageLayers extends React.Component<IManageLayersProps, any> {
             nodes: props.layers.map(li => ({ id: li.name, label: li.name, secondaryLabel: li.type, iconName: "pt-icon-layer" }))
         };
     }
-    componentWillReceiveProps(nextProps: IManageLayersProps) {
+    componentDidUpdate(prevProps: IManageLayersProps) {
+        const nextProps = this.props;
         const nodes = nextProps.layers.map(li => ({ id: li.name, label: li.name, secondaryLabel: li.type, iconName: "pt-icon-layer" }));
         this.setState({ nodes: nodes });
     }
